fix(hero part): update hero part design
