+++ src/App.tsx
@@ -1,25 +1,18 @@
 import Header from "./components/Header";
 import Hero from "./components/Hero";
-import ChainFree from "./components/ChainFree";
-import MsgportInterface from "./components/MsgportInterface";
-import EnablingDirverse from "./components/EnablingDiverse";
-import Governed from "./components/Governed";
-import Footer from "./components/Footer";
-import Resources from "./components/Resources";
-import RollingLogos from "./components/RollingLogos";
 
 function App() {
   return (
     <>
-      <Header />
-      <Hero />
-      <RollingLogos />
-      <ChainFree />
-      <MsgportInterface />
-      <EnablingDirverse />
-      <Resources />
-      <Governed />
-      <Footer />
+      <div className="relative">
+        <img
+          src="/images/cover.png"
+          className="absolute top-0 right-0 left-0 bottom-0 object-cover opacity-20"
+          alt="msport"
+        />
+        <Header />
+        <Hero />
+      </div>
     </>
   );
 }
