--- src/data/BuildMsgButtonsMobile.json
@@ -1,27 +1,0 @@
-[
-    {
-        "id" : 1 ,
-        "text" : "Deploy ExampleReceiverDapp" ,
-        "bg" : "#6F6F6F" ,
-        "selected" : "#00D448"
-    },
-    {
-        "id" : 2 ,
-        "text" : "Encode Calldata" ,
-        "bg" : "#6F6F6F" ,
-        "selected" : "#00D448"
-    },
-    {
-        "id" : 3 ,
-        "text" : "Get Fee And Params From Msgport API" ,
-        "bg" : "#6F6F6F" ,
-        "selected" : "#00D448"
-    },
-    {
-        "id" : 4 ,
-        "text" : "Sending Message" ,
-        "bg" : "#6F6F6F" ,
-        "selected" : "#00D448"
-    }
-
-]
