--- src/data/ChainFree.json
@@ -1,23 +1,0 @@
-[
-    {
-        "id" : 1 ,
-        "greenTitle" : "Chain-Free" ,
-        "whiteTitle" : "Empowerment" ,
-        "desc" : "Msgport revolutionizes DApp development, enabling developers to easily build across multiple chains, free from underlying protocol constraints, and focus on innovation and user experience" ,
-        "icon" : "/assets/icons/chainFree.svg"
-    },
-    {
-        "id" : 2 ,
-        "greenTitle" : "Composable" ,
-        "whiteTitle" : "Modules" ,
-        "desc" : "Msgport modules enhance functionality while maintaining separation from core protocols, streamlining the creation of composable, integrated DApps" ,
-        "icon" : "/assets/icons/composable.svg"
-    },
-    {
-        "id" : 3 ,
-        "greenTitle" : "Versatile" ,
-        "whiteTitle" : "Message Lines" ,
-        "desc" : "Msgport offers a unified, simple port to access a wide array of networks. It integrates a diverse range of messaging protocols like ORMP, LCMP, XCMP, and CCIP, catering to the ever-changing and varied needs of cross-chain applications" ,
-        "icon" : "/assets/icons/versatile.svg"
-    }
-]
