--- src/data/Resources.json
@@ -1,26 +1,0 @@
-[
-    {
-        "logo" : "/images/MsgportGitHub.png" ,
-        "title" : "Msgport GitHub" ,
-        "desc" : "An open-source repo providing interfaces and tools for blockchain interoperability and communication." ,
-        "link" : "https://github.com/ringecosystem/msgport"
-    } ,
-    {
-        "logo" : "/images/MessagesExplorer.png" ,
-        "title" : "Msgport Scan Explorer" ,
-        "desc" : "It serves as a comprehensive platform for following transactions over the Msgport, helping users and developers to manage and analyze cross-chain data more effectively." ,
-        "link" : "https://scan.msgport.xyz/"
-    } ,
-    {
-        "logo" : "/images/TechnicalDocs.png" ,
-        "title" : "Technical Docs" ,
-        "desc" : "Start building your Cross-Chain DApp with Msgport." ,
-        "link" : "https://msgport.ringdao.com/docs/"
-    } ,
-    {
-        "logo" : "/images/DeveloperAssistance.png" ,
-        "title" : "Developer Assistance" ,
-        "desc" : "Chat with other developers in the Developer Community, share your ideas and get answers to your technical questions." ,
-        "link" : "https://t.me/msgport_support"
-    } 
-]
