--- src/data/RollingLogo.json
@@ -1,26 +1,0 @@
-[
-    {
-        "src" : "/images/arbitrum.png" ,
-        "alt" : "arbitrum"
-    } ,
-    {
-        "src" : "/images/ethereum.png" ,
-        "alt" : "ethereum"
-    } ,
-    {
-        "src" : "/images/darwinia.png" ,
-        "alt" : "darwinia"
-    } ,
-    {
-        "src" : "/images/arbitrum.png" ,
-        "alt" : "arbitrum"
-    } ,
-    {
-        "src" : "/images/ethereum.png" ,
-        "alt" : "ethereum"
-    } ,
-    {
-        "src" : "/images/darwinia.png" ,
-        "alt" : "darwinia"
-    } 
-]
